+++ create_test/create_cache_unoptimal_bpf_program.py
@@ -8,8 +8,8 @@ def get_begin(mat_dim: int, rolled: bool):
 #define MAT_SIZE (MAT_DIM * MAT_DIM)
 int main() {{
     void *mat_map_1 = MAP_BY_FD(0), *mat_map_2 = MAP_BY_FD(1), *mat_map_res = MAP_BY_FD(2), *result;
-    int mat_1[MAT_SIZE], mat_2[MAT_SIZE], mat_res[MAT_SIZE];
-    int i, j, k;
+    float mat_1[MAT_SIZE], mat_2[MAT_SIZE], mat_res[MAT_SIZE];
+    long i, j, k;
 
 
 """
@@ -19,10 +19,10 @@ int main() {{
     // Read in 2 matrices
     for (i = 0; i < MAT_SIZE; i++) {
         result = bpf_map_lookup_elem(mat_map_1, &i);
-        mat_1[i] = *(int *)result;
+        mat_1[i] = *(float *)result;
 
         result = bpf_map_lookup_elem(mat_map_2, &i);
-        mat_2[i] = *(int *)result;
+        mat_2[i] = *(float *)result;
     }
 """
     else:
